requirements.txt
nltk
click
gitpython
